added gamelogic function
